generic forms moved fo /forms
