+++ src/dashboard/pages/dashboard/catalog/categories/edit/[id].tsx
@@ -49,7 +49,7 @@ const CategoryEditPage = () => {
     }
   }, [categoryId]);
 
-  function deleteCat() {
+  async function deleteCat() {
     deleteCategory(categoryId).then(() => {
       router.push("/dashboard/catalog/categories");
     });
