+++ src/dashboard/pages/dashboard/catalog/products/add.tsx
@@ -12,7 +12,7 @@ import TopLineWithReturn from "@/components/Dashboard/Generic/TopLineWithReturn"
 // mui
 import Container from "@mui/material/Container";
 import Grid from "@mui/material/Grid";
-import CategorySelectForm from "@/components/Dashboard/Generic/CategorySelectForm";
+import CategorySelectForm from "@/components/Dashboard/Generic/Forms/CategorySelectForm";
 import ProductBasicInfo from "@/components/Dashboard/Catalog/Products/Editor/Product/ProductBasicInfo";
 import ProductTranslatedFieldsWrapper from "@/components/Dashboard/Catalog/Products/Editor/Product/ProductTranslatedFields";
 import { Typography } from "@mui/material";
